requirements.txt
openai==1.59.3
langchain==0.3
langchain-openai==0.2.14
python-dotenv==1.0.1
ruff==0.8.1
pydantic==2.10.4
